# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Autenticación",
        "description": "API para la autenticación de usuarios",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/genotypes": {
            "get": {
                "tags": [
                    "Genotypes"
                ],
                "summary": "Obtener todos los genotipos",
                "operationId": "dd9c0c2e11b4d6faf82739469db2b08a",
                "responses": {
                    "200": {
                        "description": "Lista de genotipos"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Genotypes"
                ],
                "summary": "Crear un nuevo genotipo",
                "operationId": "2d62f54fbf4e6c9cba7894761ffd2a5a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "code"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "CCN-51"
                                    },
                                    "code": {
                                        "type": "string",
                                        "example": "GEN-001"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Genotipo de cacao resistente a enfermedades",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            },
                            "example": {
                                "name": "CCN-51",
                                "code": "GEN-001",
                                "description": "Genotipo de cacao resistente a enfermedades"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Genotipo creado exitosamente"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/genotypes/{id}": {
            "get": {
                "tags": [
                    "Genotypes"
                ],
                "summary": "Obtener un genotipo por ID",
                "operationId": "71f7bc3c71458f1b5f5e64e8b975fcc2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del genotipo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Genotipo encontrado"
                    },
                    "404": {
                        "description": "Genotipo no encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/measurements": {
            "get": {
                "tags": [
                    "Measurements"
                ],
                "summary": "Obtener todas las mediciones",
                "operationId": "67427060663f6bf2be82d6fc908f2784",
                "responses": {
                    "200": {
                        "description": "Lista de mediciones"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Measurements"
                ],
                "summary": "Crear una nueva medición",
                "operationId": "7a4ab846a45b1273acdd9e8f4695b305",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "sensor_id",
                                    "date",
                                    "data",
                                    "measurement_type"
                                ],
                                "properties": {
                                    "sensor_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "date": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2024-06-14T10:00:00"
                                    },
                                    "data": {
                                        "type": "number",
                                        "example": 23.5
                                    },
                                    "measurement_type": {
                                        "type": "string",
                                        "example": "temperatura"
                                    }
                                },
                                "type": "object"
                            },
                            "example": {
                                "sensor_id": 1,
                                "date": "2024-06-14T10:00:00",
                                "data": 23.5,
                                "measurement_type": "temperatura"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Medición creada exitosamente"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/measurements/{id}": {
            "get": {
                "tags": [
                    "Measurements"
                ],
                "summary": "Obtener una medición por ID",
                "operationId": "3aa96b516c4a271d423c91fd4a65e6be",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la medición",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Medición encontrada"
                    },
                    "404": {
                        "description": "Medición no encontrada"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Measurements"
                ],
                "summary": "Eliminar una medición por ID",
                "operationId": "2e6989a23288135e4c90128fc5d0ba33",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la medición",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Medición eliminada"
                    },
                    "404": {
                        "description": "Medición no encontrada"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}